refactor(pallet): split impls into feature-grouped impls/ module
Move the 42 Pallet<T> helper and runtime-API query methods out of the
monolithic helper impl block in lib.rs into pallet/src/impls/, grouped
by feature across 9 files (signatures, members, providers, buckets,
agreements, challenges, checkpoints, queries, marketplace), declared
from impls/mod.rs.

All methods are now pub: once they live in a sibling module, the
dispatchables in lib.rs reach them via Self::, so private/pub(crate)
visibility would no longer compile. Pure move + visibility change;
behavior is unchanged and all 194 pallet tests pass.
